requirements.txt
fastapi
uvicorn
uvicorn[standard]
langchain
langchain-openai
python-multipart
pydantic
chromadb
pytesseract
pdfminer.six
opencv-python
python-dotenv
pillow
sentence-transformers
groq
easyocr
reportlab
markdown
